Refactor to encapsulate keystrokes logic
